in Multiboot do not expect that boxmode is the latest paramter.
For more than 10 years we expect the boxmode as latest parameter in
"/sys/firmware/devicetree/base/chosen/bootargs" and the STARTUP_* files
but recently openATV added a parameter and add it at the end. See this
commit https://github.com/oe-alliance/oe-alliance-core/commit/41b655ed7ab38a3052f3e810dc18fc16541cb241
This breaks the multiboot and leads to this crash

Traceback (most recent call last):
  File "/usr/lib/enigma2/python/Components/ActionMap.py", line 76, in action
  File "/usr/lib/enigma2/python/Components/ActionMap.py", line 56, in action
  File "/usr/lib/enigma2/python/Screens/Menu.py", line 58, in okbuttonClick
  File "/usr/lib/enigma2/python/Tools/BoundFunction.py", line 10, in __call__
  File "/usr/lib/enigma2/python/Screens/Menu.py", line 71, in runScreen
  File "/usr/lib/enigma2/python/Screens/Menu.py", line 77, in openDialog
  File "/usr/lib/enigma2/python/StartEnigma.py", line 296, in openWithCallback
    dlg = self.open(screen, *arguments, **kwargs)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/enigma2/python/StartEnigma.py", line 306, in open
    dlg = self.current_dialog = self.instantiateDialog(screen, *arguments, **kwargs)
                                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/enigma2/python/StartEnigma.py", line 249, in instantiateDialog
    return self.doInstantiateDialog(screen, arguments, kwargs, self.desktop)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/enigma2/python/StartEnigma.py", line 266, in doInstantiateDialog
    dlg = screen(self, *arguments, **kwargs)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib/enigma2/python/Screens/FlashImage.py", line 514, in __init__
  File "/usr/lib/enigma2/python/Screens/FlashImage.py", line 530, in getImagesList
  File "/usr/lib/enigma2/python/Tools/Multiboot.py", line 126, in getCurrentImageMode
ValueError: invalid literal for int() with base 10: '2M'

This resolved this crash by explicitaly search for boxmode= and the
integer value.

Still for older images in the slots with these newer STARTUP_ files this
might lead to crashes. We can at least try to cherry pick this to our
older releases as well and/or openATV may also consider to relocate the
new parameter somewhere else in that string so that boxmode is again the
latest parameter.